<fix>(Disabling Plugin): More safe during disabling plugin
