+++ src/main/java/space/glowberry/fireworks/commands/commandHandler/stopLoop.java
@@ -3,6 +3,7 @@ package space.glowberry.fireworks.commands.commandHandler;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import space.glowberry.fireworks.Factory;
+import space.glowberry.fireworks.Main;
 import space.glowberry.fireworks.classes.Loop;
 import space.glowberry.fireworks.classes.LoopPool;
 import space.glowberry.fireworks.classes.LoopState;
@@ -52,7 +53,7 @@ public class stopLoop implements CommandHandler {
         }
 
         try {
-            LoopPool.getInstance().saveAllLoops();
+            Main.configManager.saveAllLoops();
         } catch (IOException e) {
             sender.sendMessage(translate(Factory.getLanguage().getString("IO-Exception")));
         }
